+++ icd-web-shared/src/main/scala/icd/web/shared/PdfOptions.scala
@@ -7,10 +7,10 @@ object PdfOptions {
   val defaultPaperSize   = "Letter"
   val defaultDetails     = true
 
-  val orientations = List("landscape", "portrait")
-  val fontSizes    = List(10, 12, 14, 16)
-  val lineHeights  = List("1.6", "1.4", "1.2", "1.0")
-  val paperSizes   = List("Letter", "Legal", "A4", "A3")
+  val orientations: List[String] = List("landscape", "portrait")
+  val fontSizes: List[Int] = List(10, 12, 14, 16)
+  val lineHeights: List[String] = List("1.6", "1.4", "1.2", "1.0")
+  val paperSizes: List[String] = List("Letter", "Legal", "A4", "A3")
 
   def apply(
       maybeOrientation: Option[String] = None,
@@ -19,7 +19,8 @@ object PdfOptions {
       maybePaperSize: Option[String] = None,
       maybeDetails: Option[Boolean] = None,
       expandedIds: List[String] = Nil,
-      processMarkdown: Boolean = true
+      processMarkdown: Boolean = true,
+      documentNumber: String = ""
   ): PdfOptions =
     PdfOptions(
       maybeOrientation.getOrElse(defaultOrientation),
@@ -28,7 +29,8 @@ object PdfOptions {
       maybePaperSize.getOrElse(defaultPaperSize),
       maybeDetails.getOrElse(defaultDetails),
       expandedIds,
-      processMarkdown
+      processMarkdown,
+      documentNumber
     )
 }
 
@@ -47,5 +49,7 @@ case class PdfOptions(
     // List of ids for expanded details
     expandedIds: List[String],
     // If true (default), process markdown in text, otherwise don't (set to false if you don't need the descriptions)
-    processMarkdown: Boolean
+    processMarkdown: Boolean,
+    // If not empty, display the string at the start of the PDF document (after the title/subtitle)
+    documentNumber: String
 )